factor out SourceFileFetcher to separate module
